+++ app/db/migrations/versions/20260112_0819_e042ea7f0764_add_reading_progress.py
@@ -5,11 +5,11 @@ Revises: 0007
 Create Date: 2026-01-12 08:19:34.831909+00:00
 
 """
-from typing import Sequence, Union
+from collections.abc import Sequence
+from typing import Union
 
-from alembic import op
 import sqlalchemy as sa
-
+from alembic import op
 
 # revision identifiers, used by Alembic.
 revision: str = 'e042ea7f0764'